Rename stats/cache to stats/svg_cache
