+++ simple/stats/svg_cache.py
