+++ simple/tests/stats/svg_cache_test.py
@@ -20,7 +20,7 @@ import unittest
 
 from parameterized import parameterized
 from proto.cache_data_pb2 import StatVarGroups
-from stats.cache import _generate_svg_cache_internal
+from stats.svg_cache import _generate_svg_cache_internal
 from stats.util import base64_decode_and_gunzip
 from stats.util import gzip_and_base64_encode
 from tests.stats.test_util import compare_files
